Split dashboard render modules
